+++ python/lsst/ts/mtdome/enums.py
@@ -19,9 +19,17 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["LlcMotionState", "LlcName", "OnOff", "ResponseCode"]
+__all__ = [
+    "LlcMotionState",
+    "LlcName",
+    "OnOff",
+    "ResponseCode",
+    "LlcNameDict",
+]
 
 import enum
+
+from lsst.ts.idl.enums.MTDome import SubSystemId
 
 
 class LlcMotionState(enum.IntEnum):
@@ -67,3 +75,7 @@ class ResponseCode(enum.IntEnum):
     OK = 0
     UNSUPPORTED_COMMAND = 2
     INCORRECT_PARAMETER = 3
+
+
+# Dictionary to look up which LlcName is associated with which sub-system.
+LlcNameDict = {getattr(SubSystemId, enum.name): enum.value for enum in LlcName}