+++ python/lsst/ts/mtdome/mock_llc/apscs.py
@@ -27,7 +27,6 @@ import numpy as np
 from lsst.ts import utils
 from .base_mock_llc import BaseMockStatus
 from ..enums import LlcMotionState
-from lsst.ts.idl.enums.MTDome import OperationalMode
 
 _NUM_SHUTTERS = 2
 _NUM_MOTORS = 4
@@ -44,7 +43,6 @@ class ApscsStatus(BaseMockStatus):
 
         # Variables holding the status of the mock Aperture Shutter
         self.status = LlcMotionState.CLOSED
-        self.operational_mode = OperationalMode.NORMAL
         self.messages = [{"code": 0, "description": "No Errors"}]
         self.position_actual = np.zeros(_NUM_SHUTTERS, dtype=float)
         self.position_commanded = 0.0
@@ -112,24 +110,6 @@ class ApscsStatus(BaseMockStatus):
         """Stop shutter motion and engage the brakes."""
         self.command_time_tai = utils.current_tai()
         self.status = LlcMotionState.STATIONARY
-
-    async def set_normal(self) -> None:
-        """Set shutter motion state to normal (as opposed to
-        degraded).
-        """
-        if self.operational_mode == OperationalMode.DEGRADED:
-            self.operational_mode = OperationalMode.NORMAL
-        else:
-            self.log.warning("Operational state already is NORMAL. Ignoring.")
-
-    async def set_degraded(self) -> None:
-        """Set shutter motion state to degraded (as opposed to
-        normal).
-        """
-        if self.operational_mode == OperationalMode.NORMAL:
-            self.operational_mode = OperationalMode.DEGRADED
-        else:
-            self.log.warning("Operational state already is DEGRADED. Ignoring.")
 
     async def exit_fault(self) -> None:
         """Clear the fault state."""