+++ doc/version_history.rst
@@ -16,7 +16,7 @@ Changes:
 * Added auto-enable capability.
 * Renamed "error" to "messages" in the status telemetry.
 * Added "operationalMode" to the status telemetry.
-* Added commands to change the operational mode of a lower level component.
+* Added commands and events to change the operational mode of a lower level component.
 * Modernized the unit tests.
 
 Requires: